+++ django_typer/config.py
@@ -3,7 +3,7 @@ import typing as t
 from django.conf import settings
 
 
-def traceback_config() -> t.Union[bool, t.Dict[str, t.Any]]:
+def traceback_config() -> t.Dict[str, t.Any]:
     """
     Fetch the rich traceback installation parameters from our settings. By default
     rich tracebacks are on with show_locals = True. If the config is set to False
@@ -12,7 +12,27 @@ def traceback_config() -> t.Union[bool, t.Dict[str, t.Any]]:
     This allows us to have a common traceback configuration for all commands. If rich
     tracebacks are managed separately this setting can also be switched off.
     """
-    cfg = getattr(settings, "DT_RICH_TRACEBACK_CONFIG", {"show_locals": True})
-    if cfg:
-        return {"show_locals": True, **(cfg if isinstance(cfg, dict) else {})}
-    return bool(cfg)
+    default = {"show_locals": False}
+    cfg = getattr(settings, "DT_RICH_TRACEBACK_CONFIG", default) or default
+    if cfg is True:
+        return default
+    return cfg
+
+
+def show_locals() -> t.Optional[bool]:
+    """
+    Return the show_locals parameter from the rich traceback configuration.
+    """
+    return traceback_config().get("show_locals", None)
+
+
+def use_rich_tracebacks() -> bool:
+    """
+    Return true if rich tracebacks should be installed, False otherwise.
+    """
+    from .utils import rich_installed
+
+    cfg = getattr(settings, "DT_RICH_TRACEBACK_CONFIG", {"show_locals": False})
+    return rich_installed and (
+        (isinstance(cfg, dict) and not cfg.get("no_install", False)) or cfg is True
+    )