+++ tests/apps/test_app/management/commands/dj_params_inherit.py
@@ -4,7 +4,14 @@ from .dj_params_some_suppressed_init import Command as BaseCommand
 
 
 class Command(BaseCommand):
-    suppressed_base_arguments = {"verbosity", "skip_checks", "traceback", "--no-color"}
+    suppressed_base_arguments = {
+        "verbosity",
+        "skip_checks",
+        "traceback",
+        "--no-color",
+        "--show-locals",
+        "hide_locals",
+    }
 
     tb: bool
 
@@ -27,7 +34,12 @@ class Command(BaseCommand):
         assert "skip_checks" not in ctx.params
         assert "no-color" not in ctx.params
         assert "traceback" in ctx.params
+        assert "show_locals" not in ctx.params
+        assert "hide_locals" not in ctx.params
         for param in COMMON_DEFAULTS.keys():
-            if param not in self.suppressed_base_arguments and param != "no_color":
+            if param not in self.suppressed_base_arguments and param not in [
+                "no_color",
+                "show_locals",
+            ]:
                 assert param in ctx.params
         return f"traceback={traceback}"