+++ tests/apps/test_app/management/commands/dj_params_some_suppressed_init.py
@@ -6,7 +6,13 @@ from typer.models import Context as TyperContext
 class Command(TyperCommand):
     help = "Test that django parameter suppression works as expected"
 
-    suppressed_base_arguments = {"verbosity", "skip_checks", "traceback"}
+    suppressed_base_arguments = {
+        "verbosity",
+        "skip_checks",
+        "traceback",
+        "show_locals",
+        "hide_locals",
+    }
 
     tb: bool