+++ tests/settings/settings_tb_change_defaults.py
@@ -1,3 +1,3 @@
 from .base import *
 
-DT_RICH_TRACEBACK_CONFIG = {"show_locals": False, "short": False}
+DT_RICH_TRACEBACK_CONFIG = {"show_locals": True, "short": False}