+++ tests/test_howto.py
@@ -54,6 +54,7 @@ class TestGroupsHowto(TestCase):
 │                            "/home/djangoprojects/myproject".                 │
 │                            [default: None]                                   │
 │ --traceback                Raise on CommandError exceptions                  │
+│ --show-locals              Print local variables in tracebacks.              │
 │ --no-color                 Don't colorize the command output.                │
 │ --force-color              Force colorization of the command output.         │
 │ --skip-checks              Skip system checks.                               │
@@ -175,6 +176,8 @@ class TestDefaultOptionsHowto(TestCase):
 │                                               [default: None]                │
 │ --traceback                                   Raise on CommandError          │
 │                                               exceptions                     │
+│ --show-locals                                 Print local variables in       │
+│                                               tracebacks.                    │
 │ --no-color                                    Don't colorize the command     │
 │                                               output.                        │
 │ --force-color                                 Force colorization of the      │